Product Category In Admin Panel Completed

+++ TajhizMining_NewVersion/01_Framework/Domain/EntityBase.cs
@@ -2,7 +2,7 @@
 {
     public class EntityBase
     {
-        public int Id { get; private set; }
+        public long Id { get; private set; }
         public DateTime CretionDate { get; set; }
 
         public EntityBase()
